+++ CHANGELOG.md
@@ -98,6 +98,11 @@
   [nandhinisubbu](https://github.com/nandhinisubbu)
   [#6304](https://github.com/realm/SwiftLint/issues/6304)
 
+* Add `ignores_literals` configuration for `trailing_whitespace` rule.
+  It allows to ignore literals.  
+  [nandhinisubbu](https://github.com/nandhinisubbu)
+  [#6194](https://github.com/realm/SwiftLint/issues/6194)
+
 ### Bug Fixes
 
 * Ignore function, initializer and subscript declarations alike when the